Fix layout tests in simulator rdar://121324669 https://bugs.webkit.org/show_bug.cgi?id=267822
Reviewed by Brent Fulgham.

Fix an issue where the Simulator entitlements are not actually applied. Also use the same bundle ID
for WebKit extensions on Simulator as on device. Additionally, the test runner app needs an extra
entitlement for launching extensions to succeed. This patch changes the bundle IDs for the XPC
services on Simulator where extension based launch is enabled. Otherwise, these bundle IDs will
conflict with the extension bundle IDs.

* Source/WebKit/Configurations/BaseExtension.xcconfig:
* Source/WebKit/Configurations/BaseXPCService.xcconfig:
* Source/WebKit/Configurations/GPUExtension.xcconfig:
* Source/WebKit/Configurations/GPUService.xcconfig:
* Source/WebKit/Configurations/NetworkingExtension.xcconfig:
* Source/WebKit/Configurations/NetworkService.xcconfig:
* Source/WebKit/Configurations/WebContentCaptivePortalExtension.xcconfig:
* Source/WebKit/Configurations/WebContentService.xcconfig:
* Source/WebKit/Configurations/WebContentExtension.xcconfig:
* Source/WebKit/Scripts/process-entitlements.sh:
* Source/WebKit/UIProcess/Launcher/cocoa/ProcessLauncherCocoa.mm:
(WebKit::ProcessLauncher::launchProcess):
* Source/WebKit/UIProcess/WebProcessPool.cpp:
(WebKit::WebProcessPool::createNewWebProcess):
* Tools/WebKitTestRunner/Configurations/WebKitTestRunnerApp-iOS.entitlements:

Canonical link: https://commits.webkit.org/273374@main

+++ Source/WebKit/Scripts/process-entitlements.sh
@@ -495,7 +495,7 @@ plistbuddy Clear dict
 # Simulator entitlements should be added to Resources/ios/XPCService-ios-simulator.entitlements
 if [[ "${WK_PLATFORM_NAME}" =~ .*simulator ]]
 then
-    if [[ "${PRODUCT_NAME}" != adattributiond && "${PRODUCT_NAME}" != webpushd ]]; then
+    if [[ -e "${CODE_SIGN_ENTITLEMENTS}" && "${PRODUCT_NAME}" != adattributiond && "${PRODUCT_NAME}" != webpushd ]]; then
         cp "${CODE_SIGN_ENTITLEMENTS}" "${WK_PROCESSED_XCENT_FILE}"
     fi
 elif [[ "${WK_PLATFORM_NAME}" == macosx ]]
